move some tests
